+++ modules/server/src/index.ts
@@ -9,5 +9,6 @@ export { FileContainer } from './internal/UploadFile';
 export { IServiceHandler } from './interfaces';
 export * from './ServerRequestWrapper';
 export * from './ServerResponseWrapper';
-export { createBeforeDecorator } from './utils';
+export { createBeforeActionDecorator } from './utils';
 export { Status } from './enums';
+export { BeforeAction } from './BeforeAction';